extract applySorting function and showModal in routeList and Clogs
